// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/AppraisalsEnum.ts
enum AppraisalsEnum {
	BUILDING_WITH_LAND = 'building_with_land',
	LAND_ONLY = 'land_only',
	APPRAISAL_ID = 'appraisal_id',
	TYPE_MY_OWN = 'Type My Own',
	SF = 'SF',
	UNIT = 'Unit',
	BED = 'Bed',
	MULTI_FAMILY = 'multi_family',
	SALE = 'sale',
	LEASE = 'lease',
	TOPOGRAPHY = 'topography',
	LOT_SHAPE = 'lot_shape',
	UTILITIES_SELECT = 'utilities_select',
	FRONTAGE = 'frontage',
	CONDITION = 'condition',
	PARKING = 'parking',
	LAND_TYPE = 'land_type',
	DATE_SOLD = 'date_sold',
	TYPE = 'type',
	COMP_TYPE = 'comp_type',
	COMPARISON_BASIS = 'comparison_basis',
	BASEMENT = 'basement',
	EXTERIOR = 'exterior',
	ROOF = 'roof',
	ELECTRICAL = 'electrical',
	PLUMBING = 'plumbing',
	HEATING_COOLING = 'heating_cooling',
	WINDOWS = 'windows',
	BEDROOMS = 'bedrooms',
	BATHROOMS = 'bathrooms',
	GARAGE = 'garage',
	FENCING = 'fencing',
	FIREPLACE = 'fireplace',
	NON_SALE = 'non_sale',
	YES = 'yes',
	NO = 'no',
	PRICE_SF = '$/SF',
	PRICE_ACRE = '$/Acre',
	ACRE = 'ACRE',
	AC = 'AC',
	MOST_LIKELY_OWNER_USER = 'most_likely_owner_user',
	MAIN_STRUCTURE_BASE = 'main_structure_base',
	FOUNDATION = 'foundation',
	ADA_COMPLIANCE = 'ada_compliance',
	PROPERTY_CLASS = 'property_class',
	PROPERTY_RIGHTS = 'property_rights',
	BUSINESS_NAME = 'business_name',
	COST = 'cost',
	INCOME = 'income',
	ROADMAP = 'roadmap',
	COST_AREA_MAP = 'save-cost-area-map',
	SALE_AREA_MAP = 'save-sale-area-map',
	LEASE_AREA_MAP = 'save-lease-area-map',
	SQ_FT = 'sq_ft',
	SQUARE_FEET = 'square_feet',
	SF_SOURCE = 'sf_source',
	ORDER_BY_COLUMN = 'order',
	ID = 'id',
	NA = 'N/A',
	PERCENT = '%',
	LAND_ONLY_TYPE = 'Land Only',
	BUILDING_SIZE = 'building_size',
	CALCULATE_LAND = 43560,
	STREET_ADDRESS = 'street_address',
	PENDING = 'Pending',
	LAND = 'land',
	INCOME_COMPARISON = 'income_comparison',
	RENT_ROLL = 'rent_roll',
}