// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/CompsEnum.ts
enum CompsEnum {
	BUILDING_WITH_LAND = 'building_with_land',
	LAND_ONLY = 'land_only',
	COMP_ID = 'comp_id',
	TYPE_MY_OWN = 'Type My Own',
	LEASE = 'lease',
	SF = 'SF',
	UNIT = 'Unit',
	BED = 'Bed',
	MULTI_FAMILY = 'multi_family',
	SALE = 'sale',
	TOPOGRAPHY = 'topography',
	LOT_SHAPE = 'lot_shape',
	UTILITIES_SELECT = 'utilities_select',
	FRONTAGE = 'frontage',
	CONDITION = 'condition',
	PARKING = 'parking',
	LAND_TYPE = 'land_type',
	DATE_SOLD = 'date_sold',
	TYPE = 'type',
	COMP_TYPE = 'comp_type',
	COMPARISON_BASIS = 'comparison_basis',
	OFFEROR_TYPE = 'offeror_type',
	ACQUIRER_TYPE = 'acquirer_type',
	BASEMENT = 'basement',
	EXTERIOR = 'exterior',
	ROOF = 'roof',
	ELECTRICAL = 'electrical',
	PLUMBING = 'plumbing',
	HEATING_COOLING = 'heating_cooling',
	WINDOWS = 'windows',
	BEDROOMS = 'bedrooms',
	BATHROOMS = 'bathrooms',
	GARAGE = 'garage',
	FENCING = 'fencing',
	FIREPLACE = 'fireplace',
	RES_COMP_ID = 'res_comp_id',
	ANNUAL = 'annual',
	MONTHLY = 'monthly',
	ZONE = 'zone',
	TOPOGRAPHIES = 'topographies',
	MULTIFAMILY = 'Multi-Family',
	FRONTAGES = 'frontages',
	PERCENT = '%',
	ACRE = 'ACRE',
	AC = 'AC',
	NON_SALE = 'non_sale',
	YES = 'yes',
	NO = 'no',
}

export default CompsEnum;